Extract webSuite from apiserver_test
Extract the `webSuite` type from apiserver_test.go and bring it into the
lib/web package so that it can be used from outside the lib/web package.
This will allow the movement of the enterprise SSO code to the
enterprise repository, retaining the existing tests.

`webSuite` has been renamed to `TestWebSuite` to highlight that it
exists for testing purposes and is not part of the lib/web package for
non-testing purposes. Ideally this would be moved to a separate fixtures
package but circular dependencies prevent that for now - larger changes
would be needed.

Some ancillary types and functions have also been extracted as they were
needed by `TestWebSuite`.
